addressing ellmaster's code comments.  removed caches and moved Lesson3 class to own file.
